pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telescope-scheduler"
version = "0.1.0"
description = "OpenEnv environment for autonomous exoplanet observation scheduling at Mauna Kea Observatory"
requires-python = ">=3.11"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2.0",
    "pandas",
    "numpy",
    "astropy",
    "openai",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = ["pytest", "httpx"]
